Strip out BLE semaphore
Remove BLE 'OK' state semaphore as it wouldn't being accessed by other
threads. On the other hand, because sdGetTimeout is a blocking read,
ble_command_withOK() will always get timed out.

The proper way of doing this is to create a task for each uart tx
transmission, and clear/join the task when receiving an OK on
first-come-first-serve basis.

v1.4 code is merged into v1.5 as they share the same pin configuration.